AddAppointmentImprovments

+++ src/main/java/com/example/medontes5/DAL/AppointmentRepository.java
@@ -21,6 +21,6 @@ public interface AppointmentRepository extends JpaRepository<Appointment, Intege
     Appointment cancelAppointmentByTime(int doc , Date data, int pat);
 
     @Query("select c from Appointment c where c.doctorEntity.doctorPersonalNumber=?1 and c.dateAndTime=?2  and c.time=?3")
-    List<Appointment> findAppointmentByT(int doc , Date data , double ora);
+    List<Appointment> findAppointmentByT(int doc , Date data , float ora);
 
 }
