+++ src/main/java/com/example/medontes5/Services/IAppointmentService.java
@@ -12,13 +12,13 @@ public interface IAppointmentService {
     void addNewFreeAppointments(Appointment a);
     DoctorEntity getDoctorByPrNumber(int nrPersonal);
     PatientEntity getPatientByPrNumber(int nrPersonal);
-    void editAppointment(int doc , Date data , PatientEntity p, double time);
+    void editAppointment(int doc , Date data , PatientEntity p, float time);
     List<Appointment> getAppByDoc(int doc);
     List<Appointment> getAppByPat(int pat);
     void cancelAppointment(int doc , Date data,int pat);
     void deleteAppointment(int doc , Date date);
     List<Appointment> todayAppDoc(int doc);
     List<Appointment> todayAppPat(int pat);
-    List<Appointment> byTime(int idD , Date data , double time);
+    List<Appointment> byTime(int idD , Date data , float time);
 
 }
